# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/current": {
      "get": {
        "operationId": "AuthController_currentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh-access-token": {
      "post": {
        "operationId": "AuthController_refreshAccessToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshAccessTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/user/signup": {
      "post": {
        "operationId": "sign-up",
        "summary": "Signup",
        "description": "Signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/user/info": {
      "put": {
        "operationId": "update",
        "summary": "Update user infomation",
        "description": "Update user infomation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/v1/user/password": {
      "put": {
        "operationId": "change-password",
        "summary": "Change password",
        "description": "Change password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePassWordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "base-nestjs-mongo API spec",
    "description": "Base nestjs mongo | [swagger.json](swagger.json) | [swagger-2.0.json](swagger-2.0.json)",
    "version": "0.0.1",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "access-token": {
        "type": "apiKey",
        "in": "header",
        "name": "access-token"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456@Abc"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshAccessTokenDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "123456@Abc"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          }
        },
        "required": [
          "address",
          "first_name",
          "last_name"
        ]
      },
      "UpdatePassWordDto": {
        "type": "object",
        "properties": {
          "currently_pass": {
            "type": "string"
          },
          "new_pass": {
            "type": "string"
          }
        },
        "required": [
          "currently_pass",
          "new_pass"
        ]
      }
    }
  }
}